# api/infra/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/review/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all reviewsn",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Get all reviews",
                "operationId": "get-review",
                "responses": {
                    "201": {
                        "description": "Successfully created review",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reviewdto.ReviewOutputDTO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/review/{review_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a review with the provided ID",
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "operationId": "delete-review",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Review ID to be delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Error details",
                        "schema": {
                            "$ref": "#/definitions/validation_errors.HTTPErrorCar"
                        }
                    }
                }
            }
        },
...
        "/api/v1/specification/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new specification with the provided information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specification"
                ],
                "summary": "Create a new specification",
                "operationId": "post-specification",
                "parameters": [
                    {
                        "description": "Specification information to be created",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/specificationdtos.SpecificationInputDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created specification",
                        "schema": {
                            "$ref": "#/definitions/specificationdtos.SpecificationOutputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/validation_errors.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/orders/{carID}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new order with the given parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Car ID",
                        "name": "carID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order input data",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/orderdto.OrderInputPartialDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created order",
                        "schema": {
                            "$ref": "#/definitions/orderdto.OrderOutputDTO"
                        }
                    }
                }
            }
        },
        "/orders/{order_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an order by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Delete order by ID",
                "operationId": "delete-order-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "other error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve user details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdtos.UserOutPutDTO"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update user details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userdtos.UserUpdateDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdtos.UserUpdateDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "orderdto.OrderInputPartialDTO": {
            "type": "object",
            "required": [
                "order_status",
                "rental_end_date",
                "rental_start_date",
                "total_cost"
            ],
            "properties": {
                "order_status": {
                    "type": "boolean"
                },
                "rental_end_date": {
                    "type": "string"
                },
                "rental_start_date": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                }
            }
        },
        "orderdto.OrderOutputDTO": {
            "type": "object",
            "properties": {
                "car_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_status": {
                    "type": "boolean"
                },
                "rental_end_date": {
                    "type": "string"
                },
                "rental_start_date": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "reviewdto.ReviewOutputDTO": {
            "type": "object",
            "properties": {
                "carId": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "specificationdtos.SpecificationInputDto": {
            "type": "object",
            "required": [
                "car_id",
                "description",
                "name"
            ],
            "properties": {
                "car_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "specificationdtos.SpecificationOutputDto": {
            "type": "object",
            "properties": {
                "car_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "userdtos.UserOutPutDTO": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "userdtos.UserUpdateDTO": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "validation_errors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "validation_errors.HTTPErrorCar": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 422
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['name is required'",
                        " 'description is required'",
                        " 'licenseplate is required'",
                        " 'brand is required'",
                        " 'A car needs a category to be registered']"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}